% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{census_reverse_geocode}
\alias{census_reverse_geocode}
\title{Reverse geocode with census}
\usage{
census_reverse_geocode(lon, lat, ...)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{...}{one or more arguments}
}
\description{
Reverse geocode with census
}
