% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_address}
\alias{geocode_address}
\title{Geocode an address.}
\usage{
geocode_address(address, key = NULL)
}
\arguments{
\item{address}{A string with the address location}

\item{key}{An API key for the Google Geocoding API. If NULL,
the function looks for a 'GOOGLE_API_KEY' environment variable.}
}
\value{
A list with elements:
\itemize{
\item address
\item cbsa_id
\item cbsa_name
\item census_tract_id
\item census_tract_name
\item county_id
\item county_name
\item latitude
\item longitude
\item state_abbr
\item state_id
\item state_name
\item zip
}

cbsa_id, census_tract_id, county_id, state_id are all FIPS codes.
}
\description{
Geocode an address.
}
\details{
This function first uses the Google Geocoding API to
get lat/lon coordinate for the address, the reverse geocodes those
coordinates with the Census geocoding API to find the CBSA and census
tract, and fill in any other information missing from the Google
geocoder response.
}
\examples{
\dontrun{
hq <- geocode_address('161 Avenue of the Americas New York, NY 10013')
}

}
