% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{google_geocode_address}
\alias{google_geocode_address}
\title{Geocode Address with Google}
\usage{
google_geocode_address(address, key = NULL, ...)
}
\arguments{
\item{address}{A street address}

\item{key}{API key}

\item{...}{one or more arguments}
}
\description{
Geocode Address with Google
}
